chore: general refactoring and adding comments
